+++ App/capstone21/capstone21/Feat/Menu/ViewModel/MenuImagePickerViewModel.swift
@@ -47,7 +47,8 @@ public class MenuImagePickerViewModel: ObservableObject {
             navigationRouter.pop()
             
         case .nextButtonDidTap:
-            Providers.HomeProvider.request(target: .postMenuImage, instance: BaseResponse<ImageResult>.self) { [weak self] data in
+            guard let image = selectedImage?.jpegData(compressionQuality: 1.0) else { return }
+            Providers.HomeProvider.request(target: .postMenuImage(image), instance: BaseResponse<ImageResult>.self) { [weak self] data in
                 if data.success {
                     guard let data = data.data else { return }
                     self?.navigationRouter.push(to: .menuAnalysisLoading(data.url))
