+++ App/capstone21/capstone21/Feat/Onboarding/ViewModel/OnboardingViewModel.swift
@@ -12,7 +12,6 @@ public class OnboardingViewModel: ObservableObject {
     
     enum Action {
         case startButtonDidTap
-        case alreadyRegisteredButtonDidTap
     }
     
     @Published var index: Int = 0
@@ -25,8 +24,6 @@ public class OnboardingViewModel: ObservableObject {
     func send(_ action: Action) {
         switch action {
         case .startButtonDidTap:
-            navigationRouter.push(to: .logIn)
-        case .alreadyRegisteredButtonDidTap:
             navigationRouter.push(to: .logIn)
         }
     }
