+++ App/capstone21/capstone21/Networks/Foundation/MoyaLoggerPlugin.swift
@@ -1,71 +1,72 @@
 //
 //  MoyaLoggerPlugin.swift
-//  HMH_iOS
+//  ZOOC
 //
-//  Created by 지희의 MAC on 1/11/24.
+//  Created by 장석우 on 2022/12/31.
 //
 
-import Foundation
+import UIKit
 
-import Foundation
 import Moya
 
 final class MoyaLoggingPlugin: PluginType {
-  // Request를 보낼 때 호출
-  func willSend(_ request: RequestType, target: TargetType) {
-    guard let httpRequest = request.request else {
-      print("❌-->유효하지 않은 요청")
-      return
+
+    // MARK: - Request 보낼 시 호출
+    func willSend(_ request: RequestType, target: TargetType) {
+        guard let httpRequest = request.request else {
+            print("--> 유효하지 않은 요청")
+            return
+        }
+        let url = httpRequest.description
+        let method = httpRequest.httpMethod ?? "메소드값이 nil입니다."
+        var log = "----------------------------------------------------\n1️⃣[\(method)] \(url)\n----------------------------------------------------\n"
+        log.append("2️⃣API: \(target)\n")
+        if let headers = httpRequest.allHTTPHeaderFields, !headers.isEmpty {
+            log.append("header: \(headers)\n")
+        }
+        if let body = httpRequest.httpBody, let bodyString = String(bytes: body, encoding: String.Encoding.utf8) {
+            log.append("\(bodyString)\n")
+        }
+        log.append("------------------- END \(method) -------------------")
+        print(log)
     }
-    let url = httpRequest.description
-    let method = httpRequest.httpMethod ?? "unknown method"
-    var log = "✨✨✨✨✨✨✨✨✨✨✨✨✨✨\n\n[\(method)] \(url)\n\n✨✨✨✨✨✨✨✨✨✨✨✨✨✨\n"
-    log.append("API: \(target)\n")
-    if let headers = httpRequest.allHTTPHeaderFields, !headers.isEmpty {
-      log.append("header: \(headers)\n")
+
+    // MARK: - Response 받을 시 호출
+    func didReceive(_ result: Result<Response, MoyaError>, target: TargetType) {
+        switch result {
+        case let .success(response):
+            self.onSucceed(response, target: target)
+        case let .failure(error):
+            self.onFail(error, target: target)
+        }
     }
-    if let body = httpRequest.httpBody, let bodyString = String(bytes: body, encoding: String.Encoding.utf8) {
-      log.append("\(bodyString)\n")
+
+    func onSucceed(_ response: Response, target: TargetType) {
+        let request = response.request
+        let url = request?.url?.absoluteString ?? "nil"
+        let statusCode = response.statusCode
+        var log = "------------------- Reponse가 도착했습니다. -------------------"
+        log.append("\n3️⃣[\(statusCode)] \(url)\n")
+        log.append("API: \(target)\n")
+        log.append("Status Code: [\(statusCode)]\n")
+        log.append("URL: \(url)\n")
+        log.append("response: \n")
+        if let reString = String(bytes: response.data, encoding: String.Encoding.utf8) {
+            log.append("4️⃣\(reString)\n")
+        }
+        log.append("------------------- END HTTP -------------------")
+        print(log)
     }
-    log.append("✨✨✨✨✨✨✨✨ END \(method) ✨✨✨✨✨✨✨✨")
-    print(log)
-  }
-  // Response가 왔을 때
-  func didReceive(_ result: Result<Response, MoyaError>, target: TargetType) {
-    switch result {
-    case let .success(response):
-      onSuceed(response, target: target, isFromError: false)
-    case let .failure(error):
-      onFail(error, target: target)
+
+    func onFail(_ error: MoyaError, target: TargetType) {
+        if let response = error.response {
+            onSucceed(response, target: target)
+            return
+        }
+        var log = "네트워크 오류"
+        log.append("<-- \(error.errorCode)\n")
+        log.append("\(error.failureReason ?? error.errorDescription ?? "unknown error")\n")
+        log.append("<-- END HTTP")
+        print(log)
     }
-  }
-    
-  func onSuceed(_ response: Response, target: TargetType, isFromError: Bool) {
-    let request = response.request
-    let url = request?.url?.absoluteString ?? "nil"
-    let statusCode = response.statusCode
-    var log = "✅✅✅✅✅ 네트워크 통신 성공 ✅✅✅✅✅"
-    log.append("\n[\(statusCode)] \(url)\n----------------------------------------------------\n")
-    log.append("API: \(target)\n")
-    response.response?.allHeaderFields.forEach {
-      log.append("\($0): \($1)\n")
-    }
-    if let reString = String(bytes: response.data, encoding: String.Encoding.utf8) {
-      log.append("\(reString)\n")
-    }
-    log.append("✅✅✅✅✅ END HTTP (\(response.data.count)-byte body) ✅✅✅✅✅")
-    print(log)
-  }
-    
-  func onFail(_ error: MoyaError, target: TargetType) {
-    if let response = error.response {
-      onSuceed(response, target: target, isFromError: true)
-      return
-    }
-    var log = "네트워크 오류"
-    log.append("😭 <-- \(error.errorCode) \(target)\n")
-    log.append("\(error.failureReason ?? error.errorDescription ?? "unknown error")\n")
-    log.append("😭 <-- END HTTP")
-    print(log)
-  }
 }
